# Conflicts:
#	app/src/main/java/com/twlrg/twsl/fragment/HomeFragment.java
#	app/src/main/java/com/twlrg/twsl/fragment/OrderListFragment.java
#	app/src/main/res/layout/item_hotel.xml

+++ app/src/main/java/com/twlrg/twsl/entity/HotelInfo.java
@@ -14,7 +14,7 @@ public class HotelInfo implements Serializable
     private String id;//4   //酒店ID
     private String title;//"某某酒店1  //酒店名称
     private String jl;//12472.47   //当前位置距离酒店距离
-    private double price;//200    //无早最低价格
+    private int price;//200    //无早最低价格
     private String position_label;//"某某地铁/某某商圈   //位置标签
     private String service_label;//"前台热情/会议酒店" //特色服务标签
     private String  reviews_label;
@@ -61,7 +61,7 @@ public class HotelInfo implements Serializable
         this.id = obj.optString("id");
         this.title = obj.optString("title");
         this.jl = obj.optString("jl");
-        this.price = obj.optDouble("price", 0);
+        this.price = obj.optInt("price", 0);
         this.position_label = obj.optString("position_label");
         this.service_label = obj.optString("service_label");
         this.lng = obj.optString("lng");
@@ -155,12 +155,12 @@ public class HotelInfo implements Serializable
         this.service_label = service_label;
     }
 
-    public double getPrice()
+    public int getPrice()
     {
         return price;
     }
 
-    public void setPrice(double price)
+    public void setPrice(int price)
     {
         this.price = price;
     }
@@ -514,4 +514,5 @@ public class HotelInfo implements Serializable
     {
         this.reviews_label = reviews_label;
     }
+
 }
